refactor: Enhance type safety and code clarity in multiple files
- Improved type casting in ListRequestData to ensure pagination values are always integers.
- Updated method signatures in ModelMakeCommand and other classes to enhance clarity and maintainability.
- Refactored regex patterns in various files for consistency and improved readability.
- Streamlined pagination logic in Grid and Entity classes for better performance and clarity.
- Introduced new helper classes for managing closure tables and tree collections, enhancing modularity and reusability.
- Improved error handling and logging practices in search jobs for better maintainability.

+++ app/Casts/ListRequestData.php
@@ -82,7 +82,7 @@ class ListRequestData extends SelectRequestData
 
     private function getDefaultPagination(): int
     {
-        return Setting::query()->where('name', 'pagination')->first('value')?->value ?? 25;
+        return (int) Setting::query()->where('name', 'pagination')->first('value')?->value ?? 25;
     }
 
     public function calculateTotalPages(int $totalRecords): int
